backend/ai-scin-deployment-20250731_050111/requirements.txt
flask==3.1.1
flask-cors==3.0.10
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0

# AI SCIN Deployment: Core AI dependencies
numpy==1.24.3
opencv-python-headless==4.8.1.78
pillow==10.1.0
scikit-learn==1.3.2
pandas==2.1.4

# AI SCIN Deployment: FAISS similarity search
faiss-cpu==1.7.4

# AI SCIN Deployment: TIMM image vectorization
timm==0.9.12

# AI SCIN Deployment: Transformers for feature extraction
transformers==4.35.0
torch==2.1.0

# AI SCIN Deployment: Google Cloud Storage for SCIN dataset
gcsfs==2023.10.0
google-auth==2.23.4

# AI SCIN Deployment: Performance and monitoring
psutil==5.9.6
memory-profiler==0.61.0
